+++ idb_shim/CHANGELOG.md
@@ -1,7 +1,7 @@
-## 2.4.0-3
+## 2.4.0-5
 
-* Add new implementation based on new js_interop and web package.
-* Allow wasm compilation for web support (legacy implementation is still available for now)
+* Add new native implementation based on new js_interop and web package.
+* Allow wasm compilation for web support (legacy implementation is still available for now by importing explicitly `idb_shim_client_native_html.dart`)
 
 ## 2.3.2